backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Supabase
supabase==2.28.2

# OpenAI (GPT-4o, TTS, Whisper)
openai==1.54.4

# FFmpeg wrapper
ffmpeg-python==0.2.0

# Image generation (씬 배경)
Pillow==11.0.0

# 파일 처리
python-multipart==0.0.12
aiofiles==24.1.0

# HTTP
httpx==0.27.2
requests==2.32.3

# YouTube 업로드
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# 에러 트래킹
sentry-sdk[fastapi]==2.19.2

# 테스트
pytest==8.3.4
pytest-asyncio==0.24.0

# ZIP 패키지 생성
# (stdlib zipfile 사용, 별도 패키지 불필요)
